Shopping Cart
This is a shopping cart with three random items book, movies and balls.
It allows the customer to increase the quantity by the push of a button
or by typing in their own amount manually. There is a checkout that adds
up all the prices of each item and puts them into a total. There are
also add2cart buttons.